Remove the style setting
It made pireal look bad on my desktop, by using a theme that is not my
default theme.

Moreover, Qt has a command line option to select a different theme, so
if a user, for some reason, would really really want pireal to use a
different theme, it would be trivial for them to do so.